Set package-lock npm config to false
The default behaviour of npm install is to install the package given as
a positional arg as well as any dependency defined in package-lock.json
that is missing from node_modules. Setting `package-lock` to false
suppresses that behaviour and makes npm only install the packages
defined in the install command.